Revert "Revert "Added attributes to bgp_neighbors module (#72)" (#77)"
This reverts commit f1566742d65fe1bc4f2e29bcb404b256ef039bdd.

+++ plugins/module_utils/network/sonic/config/bgp_neighbors/bgp_neighbors.py
@@ -227,6 +227,7 @@ class Bgp_neighbors(ConfigBase):
                             new_timers['keepalive'] = keepalive
                         if holdtime is not None and holdtime != 180:
                             new_timers['holdtime'] = holdtime
+
                     if new_timers:
                         new_pg['timers'] = new_timers
                     advertisement_interval = pg.get('advertisement_interval', None)
@@ -288,10 +289,13 @@ class Bgp_neighbors(ConfigBase):
                     if timers:
                         keepalive = timers.get('keepalive', None)
                         holdtime = timers.get('holdtime', None)
+                        connect_retry = timers.get('connect_retry', None)
                         if keepalive is not None and keepalive != 60:
                             new_timers['keepalive'] = keepalive
                         if holdtime is not None and holdtime != 180:
                             new_timers['holdtime'] = holdtime
+                        if connect_retry is not None and connect_retry != 30:
+                            new_timers['connect_retry'] = connect_retry
                     if new_timers:
                         new_neighbor['timers'] = new_timers
                     advertisement_interval = neighbor.get('advertisement_interval', None)
@@ -446,21 +450,36 @@ class Bgp_neighbors(ConfigBase):
             if neighbor:
                 bgp_neighbor = {}
                 neighbor_cfg = {}
+                tmp_bfd = {}
+                tmp_ebgp = {}
                 tmp_timers = {}
                 tmp_capability = {}
                 tmp_remote = {}
+                tmp_transport = {}
                 if neighbor.get('bfd', None) is not None:
-                    bgp_neighbor.update({'enable-bfd': {'config': {'enabled': neighbor['bfd']}}})
+                    if neighbor['bfd'].get('enabled', None) is not None:
+                        tmp_bfd.update({'enabled': neighbor['bfd']['enabled']})
+                    if neighbor['bfd'].get('check_failure', None) is not None:
+                        tmp_bfd.update({'check-control-plane-failure': neighbor['bfd']['check_failure']})
+                    if neighbor['bfd'].get('profile', None) is not None:
+                        tmp_bfd.update({'bfd-profile': neighbor['bfd']['profile']})
                 if neighbor.get('auth_pwd', None) is not None:
                     if (neighbor['auth_pwd'].get('pwd', None) is not None and
                             neighbor['auth_pwd'].get('encrypted', None) is not None):
                         bgp_neighbor.update({'auth-password': {'config': {'password': neighbor['auth_pwd']['pwd'],
                                                                           'encrypted': neighbor['auth_pwd']['encrypted']}}})
+                if neighbor.get('ebgp_multihop', None) is not None:
+                    if neighbor['ebgp_multihop'].get('enabled', None) is not None:
+                        tmp_ebgp.update({'enabled': neighbor['ebgp_multihop']['enabled']})
+                    if neighbor['ebgp_multihop'].get('multihop_ttl', None) is not None:
+                        tmp_ebgp.update({'multihop-ttl': neighbor['ebgp_multihop']['multihop_ttl']})
                 if neighbor.get('timers', None) is not None:
                     if neighbor['timers'].get('holdtime', None) is not None:
                         tmp_timers.update({'hold-time': neighbor['timers']['holdtime']})
                     if neighbor['timers'].get('keepalive', None) is not None:
                         tmp_timers.update({'keepalive-interval': neighbor['timers']['keepalive']})
+                    if neighbor['timers'].get('connect_retry', None) is not None:
+                        tmp_timers.update({'connect-retry': neighbor['timers']['connect_retry']})
                 if neighbor.get('capability', None) is not None:
                     if neighbor['capability'].get('dynamic', None) is not None:
                         tmp_capability.update({'capability-dynamic': neighbor['capability']['dynamic']})
@@ -475,6 +494,39 @@ class Bgp_neighbors(ConfigBase):
                     neighbor_cfg.update({'peer-group': neighbor['peer_group']})
                 if neighbor.get('nbr_description', None) is not None:
                     neighbor_cfg.update({'description': neighbor['nbr_description']})
+                if neighbor.get('disable_connected_check', None) is not None:
+                    neighbor_cfg.update({'disable-ebgp-connected-route-check': neighbor['disable_connected_check']})
+                if neighbor.get('dont_negotiate_capability', None) is not None:
+                    neighbor_cfg.update({'dont-negotiate-capability': neighbor['dont_negotiate_capability']})
+                if neighbor.get('enforce_first_as', None) is not None:
+                    neighbor_cfg.update({'enforce-first-as': neighbor['enforce_first_as']})
+                if neighbor.get('enforce_multihop', None) is not None:
+                    neighbor_cfg.update({'enforce-multihop': neighbor['enforce_multihop']})
+                if neighbor.get('override_capability', None) is not None:
+                    neighbor_cfg.update({'override-capability': neighbor['override_capability']})
+                if neighbor.get('port', None) is not None:
+                    neighbor_cfg.update({'peer-port': neighbor['port']})
+                if neighbor.get('shutdown_msg', None) is not None:
+                    neighbor_cfg.update({'shutdown-message': neighbor['shutdown_msg']})
+                if neighbor.get('solo', None) is not None:
+                    neighbor_cfg.update({'solo-peer': neighbor['solo']})
+                if neighbor.get('strict_capability_match', None) is not None:
+                    neighbor_cfg.update({'strict-capability-match': neighbor['strict_capability_match']})
+                if neighbor.get('ttl_security', None) is not None:
+                    neighbor_cfg.update({'ttl-security-hops': neighbor['ttl_security']})
+                if neighbor.get('v6only', None) is not None:
+                    neighbor_cfg.update({'openconfig-bgp-ext:v6only': neighbor['v6only']})
+                if neighbor.get('local_as', None) is not None:
+                    if neighbor['local_as'].get('as', None) is not None:
+                        neighbor_cfg.update({'local-as': neighbor['local_as']['as']})
+                    if neighbor['local_as'].get('no_prepend', None) is not None:
+                        neighbor_cfg.update({'local-as-no-prepend': neighbor['local_as']['no_prepend']})
+                    if neighbor['local_as'].get('replace_as', None) is not None:
+                        neighbor_cfg.update({'local-as-replace-as': neighbor['local_as']['replace_as']})
+                if neighbor.get('local_address', None) is not None:
+                    tmp_transport.update({'local-address': neighbor['local_address']})
+                if neighbor.get('passive', None) is not None:
+                    tmp_transport.update({'passive-mode': neighbor['passive']})
                 if neighbor.get('remote_as', None) is not None:
                     have_nei = self.find_nei(have, bgp_as, vrf_name, neighbor)
                     if neighbor['remote_as'].get('peer_as', None) is not None:
@@ -495,8 +547,14 @@ class Bgp_neighbors(ConfigBase):
                                     del_nei.update({'remote_as': have_nei['remote_as']})
                                     requests.extend(self.delete_specific_param_request(vrf_name, del_nei))
                         tmp_remote.update({'peer-type': neighbor['remote_as']['peer_type'].upper()})
+                if tmp_bfd:
+                    bgp_neighbor.update({'enable-bfd': {'config': tmp_bfd}})
+                if tmp_ebgp:
+                    bgp_neighbor.update({'ebgp-multihop': {'config': tmp_ebgp}})
                 if tmp_timers:
                     bgp_neighbor.update({'timers': {'config': tmp_timers}})
+                if tmp_transport:
+                    bgp_neighbor.update({'transport': {'config': tmp_transport}})
                 if tmp_capability:
                     neighbor_cfg.update(tmp_capability)
                 if tmp_remote:
@@ -659,6 +717,55 @@ class Bgp_neighbors(ConfigBase):
         if cmd.get('nbr_description', None) is not None:
             delete_path = delete_static_path + '/config/description'
             requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('disable_connected_check', None) is not None:
+            delete_path = delete_static_path + '/config/disable-ebgp-connected-route-check'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('dont_negotiate_capability', None) is not None:
+            delete_path = delete_static_path + '/config/dont-negotiate-capability'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('enforce_first_as', None) is not None:
+            delete_path = delete_static_path + '/config/enforce-first-as'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('enforce_multihop', None) is not None:
+            delete_path = delete_static_path + '/config/enforce-multihop'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('override_capability', None) is not None:
+            delete_path = delete_static_path + '/config/override-capability'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('port', None) is not None:
+            delete_path = delete_static_path + '/config/peer-port'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('shutdown_msg', None) is not None:
+            delete_path = delete_static_path + '/config/shutdown-message'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('solo', None) is not None:
+            delete_path = delete_static_path + '/config/solo-peer'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('strict_capability_match', None) is not None:
+            delete_path = delete_static_path + '/config/strict-capability-match'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('ttl_security', None) is not None:
+            delete_path = delete_static_path + '/config/ttl-security-hops'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('v6only', None) is not None:
+            delete_path = delete_static_path + '/config/openconfig-bgp-ext:v6only'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('local_as', None) is not None:
+            if cmd['local_as'].get('as', None) is not None:
+                delete_path = delete_static_path + '/config/local-as'
+                requests.append({'path': delete_path, 'method': DELETE})
+            if cmd['local_as'].get('no_prepend', None) is not None:
+                delete_path = delete_static_path + '/config/local-as-no-prepend'
+                requests.append({'path': delete_path, 'method': DELETE})
+            if cmd['local_as'].get('replace_as', None) is not None:
+                delete_path = delete_static_path + '/config/local-as-replace-as'
+                requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('local_address', None) is not None:
+            delete_path = delete_static_path + '/transport/config/local-address'
+            requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('passive', None) is not None:
+            delete_path = delete_static_path + '/transport/config/passive-mode'
+            requests.append({'path': delete_path, 'method': DELETE})
         if cmd.get('advertisement_interval', None) is not None:
             delete_path = delete_static_path + '/timers/config/minimum-advertisement-interval'
             requests.append({'path': delete_path, 'method': DELETE})
@@ -669,6 +776,9 @@ class Bgp_neighbors(ConfigBase):
             if cmd['timers'].get('keepalive', None) is not None:
                 delete_path = delete_static_path + '/timers/config/keepalive-interval'
                 requests.append({'path': delete_path, 'method': DELETE})
+            if cmd['timers'].get('connect_retry', None) is not None:
+                delete_path = delete_static_path + '/timers/config/connect-retry'
+                requests.append({'path': delete_path, 'method': DELETE})
         if cmd.get('capability', None) is not None:
             if cmd['capability'].get('dynamic', None) is not None:
                 delete_path = delete_static_path + '/config/capability-dynamic'
@@ -677,14 +787,28 @@ class Bgp_neighbors(ConfigBase):
                 delete_path = delete_static_path + '/config/capability-extended-nexthop'
                 requests.append({'path': delete_path, 'method': DELETE})
         if cmd.get('bfd', None) is not None:
-            delete_path = delete_static_path + '/enable-bfd/config/enabled'
-            requests.append({'path': delete_path, 'method': DELETE})
+            if cmd['bfd'].get('enabled', None) is not None:
+                delete_path = delete_static_path + '/enable-bfd/config/enabled'
+                requests.append({'path': delete_path, 'method': DELETE})
+            if cmd['bfd'].get('check_failure', None) is not None:
+                delete_path = delete_static_path + '/enable-bfd/config/check-control-plane-failure'
+                requests.append({'path': delete_path, 'method': DELETE})
+            if cmd['bfd'].get('profile', None) is not None:
+                delete_path = delete_static_path + '/enable-bfd/config/bfd-profile'
+                requests.append({'path': delete_path, 'method': DELETE})
         if cmd.get('auth_pwd', None) is not None:
             if cmd['auth_pwd'].get('pwd', None) is not None:
                 delete_path = delete_static_path + '/auth-password/config/password'
                 requests.append({'path': delete_path, 'method': DELETE})
             if cmd['auth_pwd'].get('encrypted', None) is not None:
                 delete_path = delete_static_path + '/auth-password/config/encrypted'
+                requests.append({'path': delete_path, 'method': DELETE})
+        if cmd.get('ebgp_multihop', None) is not None:
+            if cmd['ebgp_multihop'].get('enabled', None) is not None:
+                delete_path = delete_static_path + '/ebgp-multihop/config/enabled'
+                requests.append({'path': delete_path, 'method': DELETE})
+            if cmd['ebgp_multihop'].get('multihop_ttl', None) is not None:
+                delete_path = delete_static_path + '/ebgp-multihop/config/multihop_ttl'
                 requests.append({'path': delete_path, 'method': DELETE})
 
         return requests
